Add a random component to the optimization algorithm.
The previous method simply iterated through each channel of each color,
and chose the value that most improved the overall result. This was
okay, but it did occasionally tend to get stuck in local optima, as it
was unable to change more than one channel in any given step.

This commit instead alternates between this existing method and a random
method that tries several randomly generated colors for each palette
entry and checks to see if it improves the situation.

This is intended to be a stopgap measure until a better overall
algorithm can be determined.

Signed-off-by: Jason Lynch <jason@calindora.com>