Change the error function to be based on SSIM.
The existing error function simply added up the errors from individual
pixels. This resulted in dithering being improperly evaluated and some
pathological cases could produce highly unintuitive results. The error
is now computed using a variant of the SSIM (structural similarity
index measure) algorithm provided by DSSIM.

It was initially an experiment, but the improvement in the final results
were impressive enough to warrant keeping it. It's possible there is a
more applicable metric that could also be used.

Signed-off-by: Jason Lynch <jason@calindora.com>